Begun module restructuring
