+++ src/app/user/activities/activities.component.html
