+++ src/app/user/activities/activities.component.scss
